backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
pgvector==0.2.4

# Authentication
djangorestframework-simplejwt==5.3.1

# Async & Task Queue
celery==5.3.6
redis==5.0.1
django-redis==5.4.0

# ASGI Server
uvicorn==0.25.0
uvicorn[standard]==0.25.0

# HTTP & WebSocket
aiohttp==3.9.1
websockets==12.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Environment
python-decouple==3.8
python-dotenv==1.0.0

# Utilities
python-slugify==8.0.1
whitenoise==6.6.0
orjson==3.9.10  # Fast JSON serialization for improved streaming performance

# Development
black==23.12.1
ruff==0.1.9
mypy==1.8.0
django-stubs==4.2.7
djangorestframework-stubs==3.14.5

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.23.2
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==21.0.0

# Monitoring
sentry-sdk==1.39.2

# Payment (for later)
stripe==7.9.0
